Api - add required changes

+++ app/controllers/api/v1/motorcycles_controller.rb
@@ -35,27 +35,34 @@ class Api::V1::MotorcyclesController < ApplicationController
 
   # UPDATE MOTORCYCLE
   def update
-    @motorcycle = Motorcycle.find(params[:id])
-    update_cycle = params.require(:motorcycle)
-      .permit(:cylinder, :description, :model, :acceleration, :title, :price, :duration, :discount, images: [])
-      .merge(user: @user)
-    if @motorcycle.update(update_cycle)
-      render json: { motorcycle: @motorcycle, message: 'Motorcycle updated successfully!' }
+    if @motorcycle = @user.motorcycle.find_by(id: params[:id])
+      @motorcycle = Motorcycle.find(params[:id])
+      update_cycle = params.require(:motorcycle)
+        .permit(:cylinder, :description, :model, :acceleration, :title, :price, :duration, :discount, images: [])
+        .merge(user: @user)
+      if @motorcycle.update(update_cycle)
+        render json: { motorcycle: @motorcycle, message: 'Motorcycle updated successfully!' }
+      else
+        render json: { errors: @motorcycle.errors.full_messages, message: 'Motorcycle not updated!' }
+      end
     else
-      render json: { errors: @motorcycle.errors.full_messages, message: 'Motorcycle not updated!' }
+      render json: { message: 'Only the owner is permitted to update this motorcycle!' }
     end
   end
 
   # DELETE MOTORCYCLE
   def destroy
-    @motorcycle = Motorcycle.all
-    @motorcycle = Motorcycle.find(params[:id])
-    if @motorcycle.destroy
-      render json: { motorcycles: @motorcycles, message: 'Motorcycle deleted successfully!' }, status: :ok
+    if @motorcycle = @user.motorcycle.find_by(id: params[:id])
+      @motorcycle = Motorcycle.find(params[:id])
+      if @motorcycle.destroy
+        render json: { motorcycles: @motorcycles, message: 'Motorcycle deleted successfully!' }, status: :ok
+      else
+        render json: { errors: @motorcycle.errors.full_messages,
+                       message: 'Motorcycle not deleted!' },
+               status: :unauthorized_user
+      end
     else
-      render json: { errors: @motorcycle.errors.full_messages,
-                     message: 'Only the owner is permitted to delete this motorcycle!' },
-             status: :unauthorized_user
+      render json: { message: 'Only the owner is permitted to delete this motorcycle!' }
     end
   end
 
@@ -64,4 +71,8 @@ class Api::V1::MotorcyclesController < ApplicationController
   def set_motorcycle
     @motorcycle = Motorcycle.find(params[:id])
   end
+
+  # def set_permission
+
+  # end
 end
